Renamed object db actions
