+++ src/actions/object.delete-db-table.spec.ts
