+++ src/actions/object.delete-db-table.ts
@@ -1,18 +1,14 @@
 import _ from 'lodash';
-import chai from 'chai';
-import { Knex } from 'knex';
 
 // Classes
 import { AutoInstall } from '@classes/autoinstall';
 import { Action } from '@root/src/classes/action';
 import { Signal } from '@classes/signal';
 import { Record } from '@classes/record';
-import { Object } from '@classes/object';
 
 // Typedefs
 import { ActionRing } from '@root/src/typedefs/action';
 import { ObjectType } from '@typedefs/object';
-
 
 export default class extends Action {
     toName(): string {
@@ -27,7 +23,7 @@ export default class extends Action {
         return ActionRing.Post;
     }
 
-    onCreate(): boolean {
+    onDelete(): boolean {
         return true;
     }
 
@@ -36,20 +32,19 @@ export default class extends Action {
     }
 
     async one(signal: Signal, record: Record) {
-        // Setup
-        let object_name = record.data.name;
-        let object_type = record.data.type;
-        let auto = new AutoInstall(signal.kernel);
+        // let object_name = record.data.name;
+        // let object_type = record.data.type;
+        // let auto = new AutoInstall(signal.kernel);
 
-        // Only create objects that are marked as `database` types
-        if (object_type !== 'database') {
-            return;
-        }
+        // // Only delete objects that are marked as `database` types
+        // if (object_type !== 'database') {
+        //     return;
+        // }
 
-        // Create the empty table with no default columns
-        await auto.createTable(object_name, table => {});
+        // // Drop the table
+        // await auto.deleteTable(object_name);
 
-        // Add to kernel metadata
-        signal.kernel.meta.objects.set(object_name, Object.from(record.data));
+        // // Remove from kernel metadata
+        // signal.kernel.meta.objects.delete(object_name);
     }
 }
